Add .clang-format and (mostly) run clang-format
The settings in the .clang-format file aresuggestive, rather than
prescriptive.

The nature of refactoring is that it moves source code around.  Sometimes
we explicitly want a simple function spread out around multiple lines to
make sure the refactoring gets everything.

Rather than put comment barriers around all these instances, we instead
run clang-format on a file and discard changes we don't want.  It's not
perfect, but it lets clang-format catch the stupid mistakes.