Deprecate Reorder Parameter tests involving ref out parameter
This is a holdover from the original ReFactor!Pro test cases, where
reorder parameters allowed you to swap a function's return value for an
out reference parameter.  This is much more involved than just
reordering the parameters in the function, it is more akin to an
elaborate Change Signature.  Rather than remove the test cases from the
source files, put '(deprecated)' as the test description and put empty
diffs in the results.