FileContents::transform always operates on itself
The passed in path was always the path from the FileContents instance,
so just make that the normal case.  Since the rewriting of files always
operates on files under control by git, those changes can always be
reversed with git.  Therefore there is no need to operate on a different
file as an output file.